requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# File upload & static serving
python-multipart==0.0.9
aiofiles==24.1.0

# PPTX generation & manipulation
python-pptx==0.6.23
Pillow==10.4.0
lxml==5.3.0

# HTTP requests (for LLM API calls)
requests==2.32.3

# Config management & templating
python-dotenv==1.0.1
jinja2==3.1.4

# Optional (developer utilities, can remove if deploying minimal)
pydantic==2.9.2
black==24.8.0
mypy==1.11.2

python-dotenv
